+++ src/adb/core/charting/styling/__init__.py
@@ -1,4 +1,9 @@
 
 # Facade for styling
 from .theme import theme
-from .formatters import currency_formatter, percent_formatter, human_readable_formatter
+from .formatters import (
+    currency_formatter,
+    percent_formatter,
+    human_readable_formatter,
+    points_formatter,
+)